chore(icon): move scripts and SVG files from src directory
