+++ packages/icons/scripts/utils/optimize-svg.ts
