+++ packages/icons/scripts/utils/svgo.config.ts
